+++ components/button/Button.js
@@ -1,32 +1,40 @@
-import React from 'react';
+import React from "react";
 
-import './Button.less';
-import NavLink from '../nav/NavLink';
-import InlineSVG from '../InlineSVG';
+import "./Button.less";
+import NavLink from "../nav/NavLink";
+import InlineSVG from "../InlineSVG";
 
-export default function Button({url, as, className, inner, style, onClick, svg, ...props}) {
-
-    React.useEffect(() => {
-
-    }, []);
+export default function Button({
+    url,
+    as,
+    className,
+    inner,
+    style,
+    onClick,
+    svg,
+    ...props
+}) {
+    console.log("Button url", url);
+    React.useEffect(() => {}, []);
 
     const handleClick = (e) => {
-        if(onClick) {
+        if (onClick) {
             onClick(e);
         }
-    }
+    };
 
     return url ? (
         <NavLink href={url} as={as}>
             <>
-            {svg ? <InlineSVG type={svg} /> : null}
-            <a
-                className={`button${className ? ` ${className}` : ""}`}
-                style={style}
-                onClick={onClick ? handleClick : null}
-            >
-                {inner}
-            </a>
+                {svg ? <InlineSVG type={svg} /> : null}
+                <a
+                    className={`button${className ? ` ${className}` : ""}`}
+                    style={style}
+                    href={url}
+                    onClick={onClick ? handleClick : null}
+                >
+                    {inner}
+                </a>
             </>
         </NavLink>
     ) : (
@@ -38,11 +46,11 @@ export default function Button({url, as, className, inner, style, onClick, svg, 
             style={style}
         >
             <>
-            {svg ? <InlineSVG type={svg} /> : null}
-            {inner}
-            {props.type === "season" && props.current ? (
-                <span>{props.episodes}</span>
-            ) : null}
+                {svg ? <InlineSVG type={svg} /> : null}
+                {inner}
+                {props.type === "season" && props.current ? (
+                    <span>{props.episodes}</span>
+                ) : null}
             </>
         </a>
     );
