+++ components/button/Button.less
@@ -53,8 +53,8 @@
         font-size: 14px;
         font-weight: 500;
         line-height: 36px;
-        padding: 0 15px;
-        border-radius: 20px;
+        padding: 0 10px;
+        border-radius: 30px;
 
         &:before {
             content: "S";
