+++ components/channel/channel.less
@@ -1,0 +1,132 @@
+@import "../../styles/config.less";
+
+.epg-container {
+    display: flex;
+    padding-top: @headerHeight;
+
+    .epgPlayer {
+        display: flex;
+        flex-direction: column;
+        position: sticky;
+        align-self: flex-start;
+        top: @headerHeight;
+        width: 400px;
+        z-index: 90;
+        padding: 0 20px;
+        .bsBoxSizing();
+
+        .fixed-player {
+            padding-top: 70px;
+            position: relative;
+            .live-watching {
+                position: absolute;
+                top: 40px;
+                margin-left: 10px;
+                color: @colorYellow;
+                font-size: 15px Roboto;
+
+                @media (max-width: @mobileLargeBP) {
+                    display: none;
+                }
+            }
+
+            @media (max-width: @mobileLargeBP) {
+                padding-top: 0;
+                background-color: #0e212f;
+            }
+
+            &-categories {
+                margin: 10px 0px;
+                display: flex;
+
+                select {
+                    background-color: #ffffff31;
+                    opacity: 70%;
+                    border-radius: 30px;
+                    width: 100%;
+                    padding: 10px 10px;
+                    border: none;
+                    color: white;
+                    font: 15px Roboto;
+
+                    option {
+                        background-color: #132f45;
+                        border: none;
+                    }
+                }
+
+                @media (max-width: @mobileLargeBP) {
+                    display: none;
+                }
+            }
+        }
+
+        @media (max-width: @mobileLargeBP) {
+            padding: 0;
+            top: @headerHeightMob;
+            display: block;
+            width: 100%;
+        }
+        .current-channel-information {
+            display: flex;
+
+            padding: 10px 10px;
+            justify-content: space-between;
+            color: white;
+            align-items: center;
+            background-color: #132f45;
+
+            &-img {
+                float: right;
+            }
+
+            img {
+                width: 30px;
+                height: 30px;
+            }
+
+            .now-playing {
+                font: 12px Roboto;
+                color: #fbcc35;
+                text-overflow: ellipsis;
+            }
+
+            h1 {
+                align-self: center;
+                margin: 0px;
+                font: 15px Roboto;
+            }
+        }
+
+        .ads {
+            display: flex;
+            justify-content: center;
+            width: 100%;
+            margin-top: 10px;
+
+            @media (max-width: @mobileLargeBP) {
+                display: none;
+            }
+        }
+    }
+
+    @media (max-width: @mobileLargeBP) {
+        flex-direction: column;
+        padding-top: @headerHeightMob;
+    }
+}
+
+@media (min-width: (@mobileLargeBP + 1px)) {
+    .epgInactive {
+        .epgPlayer {
+            .video-js {
+                &:after {
+                    opacity: 0;
+                }
+            }
+            &:after {
+                opacity: 1;
+            }
+        }
+    }
+}
