+++ components/header/Header.less
@@ -22,9 +22,8 @@ header {
         left: 0;
         width: 100%;
         height: 100%;
-        // background: #040e19;
-        // opacity: 0.85;
-        background-color: #040e19;
+        background: #040e19;
+        opacity: 0.85;
     }
 
     &:after {
