+++ components/pdp/PDPTop.less
@@ -9,7 +9,7 @@
     position: relative;
     display: flex;
     align-items: flex-start;
-    padding: 30px @sitePadding 0;
+    padding: 40px 50px;
     margin: 0 auto;
     max-width: 1000px;
 
@@ -173,6 +173,10 @@
             align-items: center;
             justify-content: center;
             display: inline-flex;
+
+            svg {
+                margin-right: 10px;
+            }
         }
 
         & > a {
