+++ pages/movies/[...slug].js
@@ -90,12 +90,12 @@ export default function Movie({session, config, page, error, pageType, seoObj}) 
                     <PDPBg pdp={pdp}/>
                     <PDPTop pdp={pdp} type="movie" currVideo={currVideo}/>
                     { moreLikeThis ? <CardList key={`more-${pdp.id}`} type="title" data={moreLikeThis}/> : null}
-                    { genres.length > 0 ? (
+                    {/* { genres.length > 0 ? (
                         <ButtonList className="pdp-genres" type="grid" header={constants.GENRES} data={genres} />
                     ) : null}
                     { tags.length > 0 ? (
                         <ButtonList className="pdp-tags" type="grid" header={constants.TAGS} data={tags} />
-                    ) : null}
+                    ) : null} */}
                 </div>
                 <Player pageType={pageType} video={firstVideo} autostart={false} showPlayer={store.showPlayer} />
                 {/*<pre>{JSON.stringify(page, null, 2)}</pre>*/}
